third_party/blink/web_tests/wpt_internal/app-history/app-history-entry/sameDocument-after-fragment-navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // setTimeout because blink doesn't create history entries for location API
  // navigations if onload has not yet completed.
  window.onload = () => { setTimeout(t.step_func_done(() => {
    let entry1 = appHistory.current;
    assert_true(entry1.sameDocument);

    location = "#hash";
    let entry2 = appHistory.current;
    assert_not_equals(entry1, entry2);
    assert_true(entry1.sameDocument);

    history.pushState("push", "", "#push");
    let entry3 = appHistory.current;
    assert_not_equals(entry1, entry3);
    assert_not_equals(entry2, entry3);
    assert_true(entry1.sameDocument);
    assert_true(entry2.sameDocument);
  }), 0) };
}, "AppHistoryEntry.sameDocument after same-document navigations");
</script>
